finish sun and octopus

+++ light.h
@@ -6,7 +6,7 @@
 
 typedef struct _light{
 	glm::vec4 light0 = glm::vec4(10.0f, 10.0f, 100.0f, 0);
- 	GLint ifNight = 0;
+ 	GLfloat ifNight = 0;
 } LIGHT;
 
 void update_light(glm::vec3& sun_pos, LIGHT& THE_LIGHT);
